requirements.txt
Flask
python-dotenv
google-generativeai
PyPDF2
python-docx
Flask-Cors
pdf2image
pytesseract 
openai
jira
python-dotenv
PyMuPDF
google-generativeai